# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Edenlife Meal Recommendation System API V1 Documentation",
        "description": "Edenlife Meal Recommendation System API V1 Documentation",
        "contact": {
            "email": "ayodeleoniosun63@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/users/my-allergies": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show user allergies",
                "description": "Show user allergies",
                "operationId": "d8fb89fd0bef99198fe37b9a0c1da105",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "allergies": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "An error occured."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/users/allergies": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Pick one or more allergies",
                "description": "Pick one or more allergies",
                "operationId": "75764ac2f0420e3710dbefb98bc1f4fa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "allergies"
                                ],
                                "properties": {
                                    "allergies": {
                                        "type": "array",
                                        "items": {}
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "allergies": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "An error occured"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/meals": {
            "get": {
                "tags": [
                    "Meals"
                ],
                "summary": "Get list of meals",
                "description": "Returns list of meals",
                "operationId": "4d828e07be520ce5d5db07095f061c20",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "meals": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "An error occured."
                    }
                }
            }
        },
        "/meals/{id}": {
            "get": {
                "tags": [
                    "Meals"
                ],
                "summary": "Show meals and their respective allergies, main item and side items",
                "description": "Show meals and their respective allergies, main item and side items",
                "operationId": "2313b3b82528f01bc79c993587a0f2ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "meal id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "meal": {
                                            "description": "The response data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "An error occured."
                    }
                }
            }
        },
        "/users/meals/recommendations": {
            "get": {
                "tags": [
                    "Meals"
                ],
                "summary": "Show logged in user meal recommendations based on his allergy",
                "description": "Show logged in user meal recommendations based on his allergy",
                "operationId": "fec6a3228e7bd26f8ae81dbfac5b17db",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The response code",
                                            "type": "string"
                                        },
                                        "recommendations": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "500": {
                        "description": "An error occured."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}